+++ frontend/src/scenes/gui_scenes/connection_lost_scene.cpp
@@ -34,6 +34,6 @@ void connection_lost_scene::update(iscene_manager *manager, TimeDelta dt) {
     ImGui::End();
 }
 
-void connection_lost_scene::handle_message(iscene_manager *manager, uint32_t type, message *msg) {
+void connection_lost_scene::handle_message(iscene_manager *manager, uint64_t type, message *msg) {
 
 }
