+++ frontend/src/scenes/gui_scenes/main_menu_scene.h
@@ -28,6 +28,6 @@ namespace ibh {
         ~main_menu_scene() override = default;
 
         void update(iscene_manager *manager, TimeDelta dt) override;
-        void handle_message(iscene_manager *manager, uint32_t type, message* msg) override;
+        void handle_message(iscene_manager *manager, uint64_t type, message* msg) override;
     };
 }
