+++ frontend/src/scenes/gui_scenes/settings_menu_scene.cpp
@@ -66,6 +66,6 @@ void settings_menu_scene::update(iscene_manager *manager, TimeDelta dt) {
     ImGui::End();
 }
 
-void settings_menu_scene::handle_message(iscene_manager *manager, uint32_t type, message *msg) {
+void settings_menu_scene::handle_message(iscene_manager *manager, uint64_t type, message *msg) {
 
 }
